move BigNumeric test in its own file
